Acknowledge that get font height now returns float instead of int
git-svn-id: http://subversion.assembla.com/svn/Debrief@1422 cb33b658-6c9e-41a7-9690-cba343611204

+++ trunk/org.mwc.cmap.plotViewer/src/org/mwc/cmap/plotViewer/editors/chart/SWTCanvasAdapter.java
@@ -850,7 +850,7 @@ public class SWTCanvasAdapter implements CanvasType, Serializable, Editable
 			
 		{
 			FontData[] fd = _theDest.getFont().getFontData();
-			int fontHt = fd[0].height;
+			int fontHt = (int)fd[0].height;
 			// shift the y. JDK uses bottom left coordinate, SWT uses top-left
 			
 			int y2 = y - fontHt;
